+++ src/modules/user/usecase/user.delete.ts
@@ -1,4 +1,4 @@
-import UserRepository from "user/user.repository";
+import UserRepository from "../user.repository";
